Move walrus UTC DateTimeField into redash.models.types.
